requirements.txt
pyside6
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
keyring
pytest
